requirements.txt
awscli>=1.16.195
bravado>=10.4.1
requests>=2.22.0
locustio>=0.11.0
pyscard>=1.9.9
PyKCS11>=1.5.7
asn1crypto>=1.3.0
cryptography>=2.9.2
requests-toolbelt>=0.9.1
pyOpenSSL>=19.1.0
m2crypto>=0.35.2
urllib3>=1.25.8